feat: better home
