+++ packages/chain-adapters/src/cosmossdk/CosmosSdkBaseAdapter.ts
@@ -1,6 +1,11 @@
 import type { AssetId, ChainId } from '@shapeshiftoss/caip'
 import { fromChainId, generateAssetIdFromCosmosSdkDenom } from '@shapeshiftoss/caip'
-import type { Bip44Params, CosmosSdkChainId, RootBip44Params } from '@shapeshiftoss/types'
+import type {
+  Bip44Params,
+  CosmosSdkChainId,
+  RootBip44Params,
+  ThorMayaChainId,
+} from '@shapeshiftoss/types'
 import { KnownChainIds } from '@shapeshiftoss/types'
 import * as unchained from '@shapeshiftoss/unchained-client'
 import { isSome } from '@shapeshiftoss/utils'
@@ -77,6 +82,15 @@ export const cosmosSdkChainIds = [
   KnownChainIds.ThorchainMainnet,
   KnownChainIds.MayachainMainnet,
 ] as const
+
+const thorMayaChainIds = [KnownChainIds.ThorchainMainnet, KnownChainIds.MayachainMainnet] as const
+
+// is any of THORChain or THOR fork (only MAYAChain at the time of writing)
+export const isThorMayaChainId = (
+  maybeThorMayaChainId: string | ChainId,
+): maybeThorMayaChainId is ThorMayaChainId => {
+  return thorMayaChainIds.includes(maybeThorMayaChainId as ThorMayaChainId)
+}
 
 export type CosmosSdkChainAdapter = CosmosSdkBaseAdapter<CosmosSdkChainId>
